Refactor: Rename files for clarity
