--- src/pages/mainSection/mainViewTypeBtns.js
@@ -1,9 +1,0 @@
-class MainViewTypeBtns {
-  constructor() {}
-
-  createButtons() {
-    const innerHTML = `
-    
-    `
-  }
-}
