+++ src/style/main.css
@@ -1,5 +1,5 @@
 @import url('./common.css');
-@import url('./init.css');
+@import url('./reset.css');
 @import url('./variables.css');
 @import url('./header.css');
 @import url('./mainGridView.css');
